Add device functionality implemented.
Removed unused code, added the "add device" functionality so the user can add a new device to their list of devices. Changed the fetch calls in devices.js to call from the database through Node instead of using a local json file. Message added if the list of devices is empty.